fix: clean up proxy page redundancy and fix 1proxy sync empty body error
- Remove duplicate OneproxyTab from ProxyTab (was rendered in both HTTP sub-tab AND dedicated 1proxy tab)
- Move Debug Toggle and Usage Token Buffer to SystemStorageTab (Settings → General)
- Move RequestLimitsTab to Settings Advanced tab
- Fix oneproxy POST handler: tolerate empty body by checking content-type before parsing JSON
